Revert "Merge branch 'setup-saucelabs' into development"
Sauce Labs tests randomly fail. The tests complete successfully in the
browser but Sauce Labs sees them as having failed. The suspicion is that
it is related to Mocha and its reporting not working with Sauce Labs
with our current configuration. Something to be revisited at a later
point...

This reverts commit 504c2abc7ecac520d37b84a3cd8195377df26aef, reversing
changes made to e918fc7107d19100622bf5cad364ec8cea811324.